+++ scripts/run_release.jl
@@ -1,5 +1,6 @@
 using Pkg
 Pkg.activate(joinpath(@__FILE__, "../.."))
+println(".")
 
 using MarkovJunior
 markovjunior_run_tool()